@@ -3,13 +3,14 @@ package lv.gennadyyonov.hellookta.actuator;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.springframework.boot.actuate.autoconfigure.security.servlet.EndpointRequest;
-import org.springframework.boot.actuate.web.trace.servlet.HttpTraceFilter;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.web.context.SecurityContextPersistenceFilter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.security.web.util.matcher.OrRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
@@ -24,19 +25,18 @@ import static java.util.Optional.ofNullable;
 @EnableConfigurationProperties({ProxyProperties.class, SecurityProperties.class})
 @Configuration(proxyBeanMethods = false)
 @Order(50)
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+public class SecurityConfig {
 
-    private final HttpTraceFilter httpTraceFilter;
     private final ProxyProperties proxyProperties;
     private final SecurityProperties securityProperties;
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
+    @Bean
+    protected SecurityFilterChain actuatorSecurityFilterChain(HttpSecurity http) throws Exception {
         http
-            .csrf().disable()
-            .addFilterAfter(httpTraceFilter, SecurityContextPersistenceFilter.class)
-            .requestMatcher(new OrRequestMatcher(getRequestMatchers()));
+            .csrf(AbstractHttpConfigurer::disable)
+            .securityMatcher(new OrRequestMatcher(getRequestMatchers()));
         configureAuthentication(http);
+        return http.build();
     }
 
     @SneakyThrows
@@ -45,11 +45,13 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
             String[] allowedRoles = ofNullable(securityProperties.getAllowedRoles())
                 .map(roles -> roles.toArray(new String[0]))
                 .orElseThrow(() -> new IllegalArgumentException("Allowed roles MUST NOT be null!"));
-            http.httpBasic()
-                .and()
-                .authorizeRequests(requests -> requests.anyRequest().hasAnyRole(allowedRoles));
+            http
+                .authorizeHttpRequests(auth -> auth
+                    .anyRequest()
+                    .hasAnyRole(allowedRoles))
+                .httpBasic(Customizer.withDefaults());
         } else {
-            http.authorizeRequests(requests -> requests.anyRequest().permitAll());
+            http.authorizeHttpRequests(requests -> requests.anyRequest().permitAll());
         }
     }
 
