@@ -6,14 +6,16 @@ import lv.gennadyyonov.hellookta.services.AuthenticationService;
 import lv.gennadyyonov.hellookta.services.TechnicalEndpointService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
 import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
+import org.springframework.security.web.SecurityFilterChain;
 
 import static java.util.Optional.ofNullable;
 import static lv.gennadyyonov.hellookta.bff.controller.EnvironmentConfigController.ENVIRONMENT_CONFIG_SUFFIX;
@@ -21,12 +23,13 @@ import static org.apache.commons.lang3.BooleanUtils.toBoolean;
 import static org.springframework.http.HttpMethod.GET;
 import static org.springframework.http.HttpMethod.OPTIONS;
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
+import static org.springframework.security.web.util.matcher.AntPathRequestMatcher.antMatcher;
 
 @RequiredArgsConstructor
 @EnableWebSecurity
 @EnableConfigurationProperties(UserCacheProperties.class)
 @Configuration
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+public class SecurityConfig {
 
     private static final String ALL_URL_PATTERN = "/**";
 
@@ -37,29 +40,29 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     private final AuthenticationService authenticationService;
     private final UserCacheProperties userCacheProperties;
 
-    @Override
-    public void configure(HttpSecurity http) throws Exception {
-        http.sessionManagement().sessionCreationPolicy(STATELESS);
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.sessionManagement(session -> session.sessionCreationPolicy(STATELESS));
         configureCsrf(http);
         ofNullable(technicalEndpointService).ifPresent(service -> service.allowTechnicalEndpoints(http));
         http
             .headers(headersCustomizer)
-            .cors()
-            .and()
-            .authorizeRequests()
-            // Allow CORS option calls
-            .antMatchers(OPTIONS, ALL_URL_PATTERN).permitAll()
-            .antMatchers(GET, ENVIRONMENT_CONFIG_SUFFIX).permitAll()
-            .anyRequest().authenticated()
-            .and()
-            .oauth2ResourceServer().jwt();
+            .cors(Customizer.withDefaults())
+            .authorizeHttpRequests(auth -> auth
+                // Allow CORS option calls
+                .requestMatchers(antMatcher(OPTIONS, ALL_URL_PATTERN)).permitAll()
+                .requestMatchers(antMatcher(GET, ENVIRONMENT_CONFIG_SUFFIX)).permitAll()
+                .anyRequest().authenticated())
+            .oauth2ResourceServer(oauth -> oauth
+                .jwt(Customizer.withDefaults()));
+        return http.build();
     }
 
     private void configureCsrf(HttpSecurity http) throws Exception {
         if (toBoolean(csrfProperties.getCsrfEnabled())) {
             http.csrf(csrfCustomizer);
         } else {
-            http.csrf().disable();
+            http.csrf(AbstractHttpConfigurer::disable);
         }
     }
 
