@@ -1,18 +1,16 @@
 package lv.gennadyyonov.hellookta.bff.config.cors;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @ConfigurationProperties(prefix = "cors")
 @RequiredArgsConstructor
-@ConstructorBinding
 @Validated
 @Value
 public class CorsProperties {
