@@ -1,24 +1,22 @@
 package lv.gennadyyonov.hellookta.bff.config.graphql;
 
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.Valid;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 
 import static java.util.Optional.ofNullable;
 
 @ConditionalOnProperty(prefix = "graphql.task", name = "async-mode-enabled", havingValue = "true")
 @ConfigurationProperties(prefix = "graphql.task")
 @RequiredArgsConstructor
-@ConstructorBinding
 @Validated
 @Getter
 public class GraphQLTaskProperties {
