@@ -6,8 +6,6 @@ import lv.gennadyyonov.hellookta.logging.ParameterLogging;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
 
-import java.net.URI;
-
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 
 @FeignClient(name = "chuckNorrisConnector", configuration = ChuckNorrisConnectorConfig.class)
@@ -15,5 +13,5 @@ public interface ChuckNorrisConnector extends ParameterLogging {
 
     @RequestLine("GET /jokes/random")
     @Headers({CONTENT_TYPE + ": " + MediaType.APPLICATION_JSON_VALUE})
-    Joke randomJoke(URI baseUri);
+    Joke randomJoke();
 }
