@@ -1,17 +1,14 @@
 package lv.gennadyyonov.hellookta.config;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @ConfigurationProperties(prefix = "technical-endpoint")
-@ConstructorBinding
 @Validated
 @RequiredArgsConstructor
 @Value
@@ -21,9 +18,10 @@ public class TechnicalEndpointProperties {
     Boolean enabled;
     List<String> additionalRealEndpoints;
     List<String> referrerHeaderNames;
-    @Valid
+    @NotNull
     List<Endpoint> endpoints;
 
+    @Validated
     @Value
     public static class Endpoint {
 
