@@ -1,20 +1,17 @@
 package lv.gennadyyonov.hellookta.config.csrf;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import java.util.List;
 
 import static org.springframework.http.HttpMethod.GET;
 
 @ConfigurationProperties(prefix = "csrf")
-@ConstructorBinding
 @Validated
 @RequiredArgsConstructor
 @Value
@@ -27,9 +24,10 @@ public class CsrfProperties {
     @NotBlank
     String headerName;
     List<String> allowedMethods;
-    @Valid
+    @NotNull
     List<Endpoint> ignoredEndpoints;
 
+    @Validated
     @Value
     static class Endpoint {
 
