@@ -6,7 +6,6 @@ import lv.gennadyyonov.hellookta.logging.LoggingExclusion;
 import lv.gennadyyonov.hellookta.logging.ParameterLogging;
 import org.springframework.cloud.openfeign.FeignClient;
 
-import java.net.URI;
 import java.util.Map;
 
 @FeignClient(name = "userInfoConnector", configuration = UserInfoConnectorConfig.class)
@@ -14,5 +13,5 @@ public interface UserInfoConnector extends ParameterLogging {
 
     @RequestLine("GET")
     @LoggingExclusion
-    Map<String, Object> getUserInfo(URI baseUri, @LoggingExclusion @HeaderMap Map<String, Object> headers);
+    Map<String, Object> getUserInfo(@LoggingExclusion @HeaderMap Map<String, Object> headers);
 }
