@@ -1,13 +1,13 @@
 package lv.gennadyyonov.hellookta.services;
 
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import lv.gennadyyonov.hellookta.connectors.UserInfoConnector;
 import lv.gennadyyonov.hellookta.dto.UserInfo;
 import org.springframework.security.oauth2.core.oidc.StandardClaimAccessor;
 
-import java.net.URI;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
@@ -23,23 +23,14 @@ import static org.springframework.security.oauth2.core.oidc.OidcScopes.PROFILE;
 import static org.springframework.util.CollectionUtils.containsAny;
 
 @Slf4j
+@RequiredArgsConstructor
 public class UserInfoService {
 
-    private static final String USER_INFO_PATH = "/v1/userinfo";
     private static final String SCOPES_CLAIM = "scp";
     private static final Collection<String> USER_INFO_SCOPES = asList(PROFILE, EMAIL, ADDRESS, PHONE);
 
     private final AuthenticationService authenticationService;
-    private final String userInfoUri;
     private final UserInfoConnector userInfoConnector;
-
-    public UserInfoService(AuthenticationService authenticationService,
-                           String issuerUri,
-                           UserInfoConnector userInfoConnector) {
-        this.authenticationService = authenticationService;
-        this.userInfoUri = issuerUri + USER_INFO_PATH;
-        this.userInfoConnector = userInfoConnector;
-    }
 
     public UserInfo getUserInfo() {
         OktaProfile oktaProfile = getOktaProfile();
@@ -69,10 +60,9 @@ public class UserInfoService {
     private Map<String, Object> loadUserInfoAttributes(Map<String, Object> tokenAttributes) {
         Collection<String> scopes = extractCollection(tokenAttributes, SCOPES_CLAIM);
         if (containsAny(scopes, USER_INFO_SCOPES)) {
-            URI baseUri = new URI(userInfoUri);
             Map<String, Object> headers = new HashMap<>();
             headers.put(AUTHORIZATION, authenticationService.authorizationHeaderValue());
-            return userInfoConnector.getUserInfo(baseUri, headers);
+            return userInfoConnector.getUserInfo(headers);
         }
         return new HashMap<>();
     }
